backend/requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
psycopg
pydantic
python-dotenv==1.0.1
fastapi
uvicorn
neo4j
azure-storage-blob
azure.ai.formrecognizer
python-multipart
lightrag-hku













